% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{create}
\alias{create}
\title{Create a New QBit based on template}
\usage{
create(
  qbit_name,
  template = get_qbit_template("r"),
  verbose = getOption("verbose")
)
}
\arguments{
\item{qbit_name}{character; Title to be used for workspace.}

\item{template}{character; Id of template to be copied/forked from. Can be any id accessible
at quantargo, see also \url{https://www.quantargo.com/qbits/explore}}

\item{verbose}{logical; Show intermediary console outputs.}
}
\value{
Id of created workspace
}
\description{
Create a New QBit based on template
}
