% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_course}
\alias{deploy_course}
\title{Deploy Quantargo Course}
\usage{
deploy_course(
  course_id,
  path = ".",
  index = read_yaml(file.path(path, "index.yml")),
  json_files = get_json_files(path),
  asset_files = list.files(path, pattern = "[^(html)|(Rmd)|(json)|(.DS_Store)]$",
    full.names = TRUE),
  apikey = getOption("QKEY"),
  tmpdir = tempdir()
)
}
\arguments{
\item{course_id}{character; Name of the QBit function.}

\item{path}{character; Root path of course.}

\item{index}{list; Index meta data for qbit.}

\item{json_files}{character; list of json files to be uploaded through endpoint}

\item{asset_files}{character; files to be uploaded as course assets.}

\item{apikey}{character; API key from QBit platform.}

\item{tmpdir}{character; Temporary directory to be used for qbit/zip creation.}
}
\description{
After the course build process is finished the generated json files can be
uploaded through this function.
}
