% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_main_file.R
\name{empty_main_file}
\alias{empty_main_file}
\title{Create Empty Main File}
\usage{
empty_main_file(extension = "R")
}
\arguments{
\item{extension}{character; File extension to be used for main file}
}
\description{
Create an empty file named \code{main.<extension>} to be used as default for
qbit deployments.
}
