% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{get_qbit_template}
\alias{get_qbit_template}
\title{Get QBit Template}
\usage{
get_qbit_template(short_name)
}
\arguments{
\item{short_name}{character; Short name of the template.}
}
\description{
Currently the following templates are available via the \code{short_name}:
}
\details{
\describe{
\item{r}{Basic R template mapping to \href{https://www.quantargo.com/qbits/qbit-template-r-base}{qbit-template-r-base}}
\item{rmd}{Basic RMarkdown template template mapping to \href{https://www.quantargo.com/qbits/qbit-template-r-rmarkdown}{qbit-template-r-rmarkdown}}
}
}
