% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{remove}
\alias{remove}
\title{Create a New QBit based on template}
\usage{
remove(qbit_id, verbose = getOption("verbose"))
}
\arguments{
\item{qbit_id}{character; Workspace Id to be removed.}

\item{verbose}{logical; Show intermediary console outputs.}
}
\value{
New QBit id
}
\description{
Create a New QBit based on template
}
