% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{render}
\alias{render}
\title{Render RMarkdown Document}
\usage{
render(qbit_id, verbose = getOption("verbose"), ...)
}
\arguments{
\item{qbit_id}{character; Workspace Id to be removed.}

\item{verbose}{logical; Show intermediary console outputs.}

\item{...}{Additional parameters passed to \link{invoke}}
}
\value{
list including response object
}
\description{
Render RMarkdown Document
}
