% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{run}
\alias{run}
\title{Render RMarkdown Document}
\usage{
run(qbit_id, code, async = FALSE, verbose = getOption("verbose"), ...)
}
\arguments{
\item{qbit_id}{character; Workspace Id to be used for code execution.}

\item{code}{character; Code to be executed.}

\item{async}{logical; Execute code async.}

\item{verbose}{logical; Show intermediary console outputs.}

\item{...}{Additional parameters passed to \link{invoke}}
}
\value{
list including response object
}
\description{
Render RMarkdown Document
}
