% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_course.R
\name{deploy_course}
\alias{deploy_course}
\title{Deploy Quantargo Course}
\usage{
deploy_course(
  course_id,
  path = ".",
  index = read_yaml(file.path(path, "index.yml")),
  json_files = get_json_files(path),
  asset_files = get_asset_files(path),
  apikey = getOption("QKEY"),
  url = getOption("QBITURL", "https://api.quantargo.com/v2"),
  tmpdir = file.path(tempdir(), course_id)
)
}
\arguments{
\item{course_id}{character; Name of the QBit function.}

\item{path}{character; Root path of course.}

\item{index}{list; Index meta data for qbit.}

\item{json_files}{character; list of json files to be uploaded through endpoint}

\item{asset_files}{character; files to be uploaded as course assets.}

\item{apikey}{character; API key from QBit platform.}

\item{url}{character; Upload url to be used}

\item{tmpdir}{character; Temporary directory to be used for qbit/zip creation.}
}
\description{
After the course build process is finished the generated json files can be
uploaded through this function.
}
